cc/requirements.txt
Flask==3.0.3
gunicorn
Flask-JWT-Extended==4.6.0
requests==2.32.3
google-cloud-secret-manager==2.20.0
google-cloud-storage==2.16.0
tensorflow==2.16.1
pillow==10.0.1
numpy==1.26.0
selenium==4.21.0
webdriver-manager==4.0.1
firebase-admin==6.5.0